+) add nginx snippets

+++ VSCode/snippets/assitant.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 userDir=`cd && pwd`
+# VSCode 代码片段
 vscode=$userDir/.config/Code/User/snippets
 
 red='\033[0;31m'
@@ -16,15 +17,26 @@ help(){
     printf "Run：$red sh assitant.sh $green<verb> $yellow<args>$end\n"
     format="  $green%-5s $yellow%-8s$end%-20s\n"
     printf "$format" "-h" "" "help"
-    printf "$format" "-a" "" "link all *.json"
+    printf "$format" "-a" "" "link all json code-snippets files" 
     printf "$format" "-f" "file" "link specific file "
 }
+
 link_snippets(){
-	files=`ls *.json`
+	files=$(ls * | grep -E "\.(json|code-snippets)")
 	for file in $files; do
-		echo 'start link '$file
-		ln -s `pwd`/$file $vscode/$file
+		echo 'start link ' $file ' -> ' $vscode/$file
+		if [ ! -f $vscode/$file ]; then
+			ln -s `pwd`/$file $vscode/$file
+		fi
 	done
+}
+
+normalize(){
+	file=$1
+
+	while IFS= read -r -u3 line; do
+		echo '"' "$line" '",'
+	done 3< "$file"
 }
 
 case $1 in 
@@ -33,10 +45,12 @@ case $1 in
     -a)
 		link_snippets
 	;;
+	-s)
+		normalize $2
+	;;
 	-f)
 		ln -s `pwd`/$2 $vscode/$2
 	;;
 	*)
 		help ;;
 esac
-
